# [ P398 Rest API 형태의 DB 작업 완료 ]

+++ ex03/src/test/java/org/zerock/mapper/ReplyMapperTests.java
@@ -56,7 +56,7 @@ public class ReplyMapperTests {
 //	@Test
 	public void testDelete() {
 
-		int targetRno = 1;
+		Long targetRno = 1L;
 
 		mapper.delete(targetRno);
 	}
